refactor: change spacing file name to tokens
