+++ mobile/tamagui.config.ts
@@ -2,7 +2,7 @@ import { config } from "@tamagui/config/v3";
 import { themes } from "./theme/themes";
 import { createTamagui } from "tamagui";
 import { fonts } from "./theme/fonts";
-import { tokens } from "./theme/spacing";
+import { tokens } from "./theme/tokens";
 
 export const tamaguiConfig = createTamagui({
   ...config,
